Add functionality to play 5 rounds of the game
A new function called playGame was added and resulted in having to move
the original playGame function inside it and duplicated 5 times with new
humanSelection and computerSelection variables each time to store the
result.

An if-else statement was added to evaluate the player and computer
scores to determine a winner of the game--variables which were also
moved into the larger playGame function. The function was called
at the end of the script.