Rework original Roshambo game with DOM methods
Added buttons to the original HTML for human choice. For each button
there are now event listeners that will execute the playround function
and append text to the appropriate sections of the page which were
added with DOM methods.

If-else statement was used to determine a winner and end execution
with a return statement once a winner is determined.