Remove alert notification and append result
The if-else statement was modified so that it would create a paragraph
element upon completion of the game. To ensure that the paragraph is
only appeneded once, a boolean was added to the outer if-else statement
to determine if the element already exists within the document body.

The getComputerChoice function was edited to end execution after a winner
is determined as well.